Introduce Class Usage Logging
Summary:
This builds on https://pytorch.org/docs/stable/notes/large_scale_deployments.html#api-usage-logging to log when certain classes are initialized.

In this diff, I only covered bert/roberta models (and not even necessarily all of them) to prove that the system works

This will not affect OSS users, as for them, the logger is a no-op.

Reviewed By: snisarg

Differential Revision: D19744167

fbshipit-source-id: d37bd56f640e95811f9485d8d231c6985924e6d4