Introduce Class Usage Logging (#1243)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/1243

This builds on https://pytorch.org/docs/stable/notes/large_scale_deployments.html#api-usage-logging to log when certain classes are initialized.

In this diff, I only covered bert/roberta models (and not even necessarily all of them) to prove that the system works

This will not affect OSS users, as for them, the logger is a no-op.

Reviewed By: snisarg

Differential Revision: D19744167

fbshipit-source-id: f07845d1c1bd1f493c0e084661ab138b48b56ad2