add Stage.OTHERS and allow TB to print to a seperate prefix not in (TRAIN, TEST, EVAL)
Summary:
In some training algorithms, we may want to use Tensorboard to produce summary on some fine-grained training results. However, currently this is not possible because Stage num is limited to (Train, Test, Eval). Further, TensorBoard maps each Stage enum to a prefix and aggregate the metrics for that prefix.

example scenarios where we need additional stage for metrics reporting:

1. federated learning: we need to distinguish the training curve on clients, and the training curve of the global model aggregated from multiple users but evaluated on the client's training data.

2. adverserial training: we want to distinguish and report the training curve on clean data and adverserial data seperately.

3. any training algorithms that performs averaging during the training process, where we can see the difference before and after averaging.

Reviewed By: geof90, snisarg

Differential Revision: D19783880

fbshipit-source-id: ed24bf06366272dce236721e1bb381768fee5b13