FAISS Embedding Search Space [3/5]
Summary:
This diff introduces an object that surfaces an API to search an embedding space. This relies on [Facebook AI Similarity Search](https://github.com/facebookresearch/faiss), a library for search and clustering.

This implements `EmbeddingSearchSpace` which relies on a `pytext.utils.embedding.PretrainedEmbedding` object to build a search space with the following method:

* `find_nearest_neighbors`: which takes a query in the form of `List[str]` and for each item in the query searches the embedding space for nearest neighbors.

Reviewed By: geof90

Differential Revision: D19819094

fbshipit-source-id: ba063a99d5f519fb3539afc7bb11800b949c3514