Back out "Add option to pad missing label in LabelListTensorizer"
Summary:
The diff breaks unit test:
https://our.intern.facebook.com/intern/test/562949968450936?ref_report_id=2804980
Problem is in add_all method of VocabBuilder, but I don't get why we need this diff and special handling for None and empty string. This LabelListTensorizer is supposed to be list of all labels associated with an example, size could be variant for each example.

Reviewed By: hudeven

Differential Revision: D20357769

fbshipit-source-id: c2a3809709915cdde640e25c616f426e2dcc2f85