Add support for length label smoothing (#1308)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/1308

Adding support for length label smoothing, which from prior experiments improves NAR performance (reducing overfitting on the length prediction task).

This diff however runs into an interesting issue with CUDA OOM and throwing an exception here https://fburl.com/diffusion/akkh752n.

I believe this is because with label smoothing and length beam size of 5, there is a chance that max target length is the predicted length size, which can cause OOM issues. Reducing max target lengths is critical to helping with this.

Another issue is the exception being thrown. This is because of an off by one issue in conv_encoder.max_target_lengths vs conv_decoder.max_target_lengths. Instead of enforcing consistency this updates to make sure the predicted lengths (conv_encoder) are less than or equal to the (conv_decoder) ones.

Reviewed By: arbabu123

Differential Revision: D20865186

fbshipit-source-id: 3e739262b42f88a41384eb80be247730e983f081