add a new PyText get_num_examples_from_batch function in model (#1319)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/1319

Provide a get_num_examples_from_batch() function to model.

It is not enforce to be overriden, nor used often except for federated learning training setting. But it is used, pytext developers need to override them for their model. We override it for LMLSTM, DocModel, and QueryDocPairwiseRankingModel in this diff.

Reviewed By: kmalik22

Differential Revision: D21032297

fbshipit-source-id: 2086dcb81101ffbd36debf37b38debee509cf8fc