Fix Memory Issues in Metric Reporter for Classification Tasks over large Label Spaces
Summary:
Currently, we store the score for every label for every instance in the epoch for computing soft metrics and for writing these to file as part of the Test workflow. This is horrible when your label space is 10K and you have a large training/validation/test set. Simply using "num_batches_per_epoch" isnt helpful because this is not respected during valaidation or testing anyways. Also it really slows down training.

In this diff I add a is_memory_efficient flag to the Classification Metric Reporter to avoid storing all of this information when we dont really need it. A better solution would be to refactor ALL of the metric reporters and that is beyond the scope of this fix. To be honest, I'm decently happy with this compormise.

Reviewed By: snisarg

Differential Revision: D21006627

fbshipit-source-id: 20ab34ca7b10b96ff982ad8a9fb680f7811b0730