Fix "calling median on empty tensor" issue in MR
Summary:
Language model metrics reporter occassionally throws "median on empty tensor", e.g: f182865817

why?
We used to check the cross entropy tensor != 0 before calling mean/median on it, to exclude the calculation of the padding positions which is just 0, which possibly results in an empty tensor.

This diff checks the tensor posisions against pad_index instead of using 0.0:a batch can have close to 0 CE, or the predict word index happened to be pad_index.

Also a failed safe is added if the tensor[target != self.pad_index] is indeed empty.

Differential Revision: D21081502

fbshipit-source-id: 685705897bd350b1ae3af6fae1e84725e045bb70