Pass in optimizer config instead of create_optimizer to trainer
Summary:
This is diff 3 for: Change PyText FL trainer to use papaya config, instead of using super-long constructors.

Currently, we create a local optimizer **outside** fl trainer, and pass in a `Callable` to fl trainer that can `create_optimizer`.
This means that local optimizer config must be parsed **outside** the fl trainer, which is bad.

This diff changes the trainer constructor to expect a dictionary for local optimizer config. The trainer internally creates a `Callable` which is used internally

**No functional change**

Reviewed By: psuzhanhy

Differential Revision: D21059677

fbshipit-source-id: 0cbc9a408dbf27c125f1b762bd3e444477c018ee