Fix "calling median on empty tensor" issue in MR (#1322)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/1322

Language model metrics reporter occassionally throws "median on empty tensor", e.g: f182865817

why?
We used to check the cross entropy tensor != 0 before calling mean/median on it, to exclude the calculation of the padding positions which is just 0, which possibly results in an empty tensor.

This diff checks the tensor posisions against pad_index instead of using 0.0:a batch can have close to 0 CE, or the predict word index happened to be pad_index.

Also a failed safe is added if the tensor[target != self.pad_index] is indeed empty.

Reviewed By: kmalik22

Differential Revision: D21081502

fbshipit-source-id: f0382cea9b7c31dc2784e59f844d81d2a79cd1c7