Replaces deprecated integer division with floor division operator
Summary:
PyTorch deprecated integer division using the div operator in 1.5 and plans to throw a RuntimeError when this happens in 1.6. This diff updates this instance of integer division to use the floor division operator to avoid this issue.

See the same change in D21571651.

Created from Diffusion's 'Open in Editor' feature.

Reviewed By: arbabu123

Differential Revision: D21599092

fbshipit-source-id: 6d9c374322c59dfc2fb0b3d839a4e1dca55f9e5a