Better error message for misconfigured data fields
Summary:
When data field names are misconfigured, we often see an error in constructing the decoder because

1. no examples were parsed from the data because of the error
2. the label classes were not found from the empty data
3. the decoder needed the label classes to determine the last layer

Instead show a better error message when the label tensorizer did not see any examples when finding the classes, and when the decoder tries to use an empty label vocab (since the first error case may be hidden according to my run).

Also show an error when trying to compute metrics for an empty set of data. Note that empty data itself is not always a problem, e.g. training for 0 updates can be done.

Reviewed By: rutyrinott

Differential Revision: D21879758

fbshipit-source-id: 9b11e5b79e79b5bd6bc19986a2cdbe5b8befa271