sparsifier sensitivity analysis
Summary: This diff is used for sparsity sensitivity analysis for each weight tensor. User should specify the sparsifier as "SensitivityAnalysisSparsifier". Set the "pre_trained_model" path. The program will mask a specific portion of weight for each weight tensor and check what is the accuracy differences if we only prune the correspond weight tensor.

Reviewed By: mwu1993

Differential Revision: D21970938

fbshipit-source-id: 4d3390392ffe17e59f9311df3006859d467de6c2