Weight tensor selection
Summary: This diff is used for selecting the weight tensor to be pruned. After the sparsity sensitivity analysis, we know for each weight tensor, how it will effect the metric(e.g., accuracy), when compared to the baseline pre-trained model. We will skip pruning of some weight tensor if we find the metric will be effected a lot by pruned such weight tensor. We also allow the user to set the "sparsifier.config.max_skipped_weight" which limit the number of skipped weight tensor.

Reviewed By: mwu1993

Differential Revision: D22112289

fbshipit-source-id: 0b904449708e52ed0925850cbf86fce10d1e53a9