Fix gradients logging (#1402)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/1402

There was an in issue with the approach that we were using to log gradients. At the end of each epoch, we would look at the model and for each named_parameter we would log tensor.grad to tensorboard. This is incorrect becase after the last step in the epoch, the model parameters will only contain accumulated gradients for the last batch. There is no guarantee that the gradient for the last batch is representative of gradients for the entire dataset.

We fix this by logging the gradients after all optmizer steps into a variable inside the metric reporter. And at the end of each epoch, we aggregate the gradients across different batches and log into tensorboard. It's not entirely clear for me how to aggregate the gradients across different batches. We may want avg gradients over the entire dataset but since at the end of each epoch we are already aggregating over each batch its not trivial to get individual gradient for each sample and avg it later. But I dont think absolute value of gradients matter so much and hence i log the avg as well as sum of gradients. This will help us see relative differences in gradients across different part of the models. I also log the l2 norm of the gradient to make sure that whether gradient clipping was successful / needed or not.

Reviewed By: geof90

Differential Revision: D22494169

fbshipit-source-id: b2faccb0e08df86071d355c47f28d6358b02a382