Apply mask sparsifier
Summary:
This diff is applying the sparsifier mask. After each iteration of training, the trainer.run_step will invoke the sparsify function. At the first iteration, we generate the mask for the weights which should be pruned (listed in sparsifier.require_mask_parameters) since the mask is fixed. After that, each iteration, we call apply_mask() to prune the weight tensor to make sure the weight tensor are sparsified.
We introduce the analysis_state in the sparsifier, this state indicate whether this sparsifier is doing the sensitivity analysis. We don't want to pruning the weight during the analysis.

Reviewed By: mwu1993

Differential Revision: D22155733

fbshipit-source-id: 0894b50f0b5ea3d5ac64ba0f672c41889ae755da