Fixes for saving pairwise encoder modules
Summary:
When training models for pairwise classification, where the text encoder is not shared for the two texts, we currently have special logic to save both encoders. (The encoders share the same config and thus save path). This diff fixes two issues with it:

- it doesn't use PathManager, making it fail for manifold paths (the default)
- it saves the two encoders to the same file, which defeats the purpose.

Probably the best solution is to use two separate configs for the two encoders, so they can be saved via the normal module saving logic. but this seems to be least disruptive to existing configs.

Reviewed By: m3rlin45

Differential Revision: D22657250

fbshipit-source-id: c8c33685f6f7f9aa0d6843facf667cd93dc27603