Make it easy to specify a large number of labels (#1386)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/1386

For classification problems with a large label space, its not possible to explictly specify the label vocab as a list in the config (eg: megataxon where there are ~10K labels). Its also not possible to extract these from the data because there might be some labels which are missing from the training data. While this is not ideal, its not right to enforce that every label in the dataset should be present in the training data especially if the instances with these labels in the test set is quite small. These failures can be annoying and the pre-processing involved is quite cumbersome. So here I add the ability to just specify the labels as a file.

Reviewed By: snisarg

Differential Revision: D21575711

fbshipit-source-id: 263281fb5503af3dc7df04cee6d6db4ed5ce22f9