Add Selfie transformer to pytext
Summary: Add Selfie transformer to pytext for inference of selfie pretrained model. Selfie is a early fusion model that relies on transformer to fuse signals from dense features. When use_selfie=True, dense features are fed into the transformer so they can be self attended to along with other tokens, instead of concating them in the decoder MLP

Reviewed By: jingfeidu

Differential Revision: D22220238

fbshipit-source-id: b63519e12a492eea48b81e55064246f51a4b9b11