Check that random seed is set for pytext distributed training
Summary:
When random seed is not set and we're training with multiple GPUs, the randomly-initialized parameter weights from each worker are not the same. Setting random seed the same across workers fixes this.

(This also causes training to get stuck if early_stop_after is set, since workers do not agree on eval metrics and when to stop).

Reviewed By: hudeven

Differential Revision: D23817069

fbshipit-source-id: 2b4d9ece51791eb69b19340c2cc2f2c751491adb