Fix space inconsistencies with BPE (#1486)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/1486

BPE is highly sensitive to space -- given a string, such as "hello world", BPE will chunk it as "hello" and " world", then break these tokens into pieces. Instead, to solve issues with spacing in the first token, we can pass in " hello" and " world" to the BPE tokenizer to ensure each token gets encoded correctly.

Reviewed By: abhinavarora

Differential Revision: D24147358

fbshipit-source-id: 5be8e81752b6ae410d3fb07c31af352a77ea6109