Fix space inconsistencies with RoBERTaSeqLabelingTensorizer BPE
Summary: BPE is highly sensitive to space -- given a string, such as "hello world", BPE will chunk it as "hello" and " world", then break these tokens into pieces. Instead, to solve issues with spacing in the first token, we can pass in " hello" and " world" to the BPE tokenizer to ensure each token gets encoded correctly.

Differential Revision: D24147358

fbshipit-source-id: 70fe92a5a78852fb08990495db2c90670fb4e92e