simplify RobertaModel by replacing Python methods with Hydra config (#1489)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/1489

before:
we were using Python method to load models e.g. models.xlmr_base_binary_doc_classifier(pretrain=True). It holds the config in Python.

as we are embracing Hydra now, it would be duplicated to keep 2 sets config, one in Python, the other in yaml.

After:
In Python:
  models.RobertaModel(model_path=...)
With OmegaConf:
  conf = OmegaConf.load('pytext/contrib/pytext_lib/conf/task/model/xlmr_dummy.yaml')
  model = models.RobertaModel(**conf)

With Hydra:
  with initialize_config_module("pytext.contrib.pytext_lib.conf"):
    cfg = compose(
        config_name="pytext_config",
        overrides=[
            "task/model=xlmr_dummy",
        ],
    )
  model = hydra.utils.instantiate(cfg.task.model)

Reviewed By: snisarg

Differential Revision: D24231445

fbshipit-source-id: c9391ad39afb1e95f9ea29c05cf7236be7c6e748