omegaconf->native objects for torchscript (#1512)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/1512

When you try to script a module that contains a OmegaConf type:

```
Module 'XlmrTransform' has no attribute '_label_names' (This attribute exists on the Python module, but we failed to convert Python type: 'omegaconf.listconfig.ListConfig' to a TorchScript type.):
  File "/data/users/snisarg/fbsource/fbcode/buck-out/dev/gen/pytext/contrib/pytext_lib/test#binary,link-tree/pytext/contrib/pytext_lib/transforms/fb_roberta.py", line 100
    property
    def label_names(self):
        return self._label_names
```

We'll need to use native types if we want to torchscript. Escalating this to omry and maybe to see if we can do this transparently. For now, adding a util method.

This will be consumed as in the next diff that changes signature, so keeping it clean with a separate diff for discussions.

omry, can we upstream this?

Differential Revision: D24382743

fbshipit-source-id: 983ac1a0f9d7bab69047a8ee43acd89e92b53dd8