DocClassification Fluent2 Transformer
Summary:
There are still pickling issues when trying to use our transforms with onbox dataloader. Doing that inside training for now, while we investigate and work with the team to find a solution.

Sending this out for early review, we will have to add TorchText's version of SentencePiece in the predictor custom OP loading..

Reviewed By: hudeven

Differential Revision: D24462163

fbshipit-source-id: e48ee07cee01cdbe44ab8f9309f68ebec9e4656a