Update param group reset for FP16
Summary:
Certain optimizers rely on nested optimizers (such as StochasticWeightAveraging), explicitly reseting the para_group with `param_group = []` won't work in this case because the parameter groups from the nested optimizers need to be removed as well. This diff updates `Optimizer` to have a `reset_param_group` abstraction that allows each optimizer to handle resetting of param groups.

This unblocks weight averaging for FP16 training.

Reviewed By: mwu1993

Differential Revision: D24513502

fbshipit-source-id: 658eebd77783ff3e710f8a288679ae1bfeca6c8b