Add option to normalize transformer output representation
Summary:
For two-tower similarity, we use the cosine similarity between sentence embeddings. For this to work out-of-the-box when deployed, the torchscripted model should output embeddings with L2 norm = 1. (so that the final similarity will just be dot product of the embeddings). Add a flag for transformers to apply the normalization.

(currently during training, the loss function does the normalization. However only the encoder is exported for producing sentence embeddings).

Reviewed By: rutyrinott

Differential Revision: D25374546

fbshipit-source-id: 955d7b3faddda143ce3cba6d4f01f3af555b7de6