Fix FP16 Optimizers
Summary:
There was a recent fairseq diff that changed the arguments for `optimizer.step` (D25152032). This breaks FairseqFP16 Optimizer which breaks the NAR workflows as the rely on FP16 from fairseq.

For some reason Apex FP16 doesn't work with NAR yet, I am yet to figure out why. Also unclear on how we should test this, to prevent future breakage.

Reviewed By: alexeib

Differential Revision: D25432409

fbshipit-source-id: 2f50b750d115306ca869be24cd08f988432062ab