Fix FP16 Optimizers (#1548)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/1548

There was a recent fairseq diff that changed the arguments for `optimizer.step` (D25152032). This breaks FairseqFP16 Optimizer which breaks the NAR workflows as the rely on FP16 from fairseq.

For some reason Apex FP16 doesn't work with NAR yet, I am yet to figure out why. Also unclear on how we should test this, to prevent future breakage.

Reviewed By: alexeib

Differential Revision: D25432409

fbshipit-source-id: 160b69583f7ae2663a447ca1561c1a7e250b86e9