Changes to Export from export_list if present in PyTextConfig (#1553)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/1553

Added a field "export_list" to PyTextConfig.   All code was changed to replace f(pytextconfig.export) to the logical equivalent of map(f, pytextconfig.export_list) and a test was added to ensure that model_type is the same across all of the elements in the export list.

*Why not union?*
The combination of Union and List is not handled perfectly by serialize.py. (In trying to understand how serialize interacts with json, I added a few test cases. In serialize_test.py est_config_to_json_for_union_with_list_a succeeds where as est_config_to_json_for_union_with_list_b fails. The difference between the two tests is the absence/presence of wrapper "boo".  Is this the intended behavior?)

Reviewed By: mikekgfb

Differential Revision: D25257145

fbshipit-source-id: b9bc83f66ed41395518d83cd50cb46f38d2bbec0