Add lstm encoder, and models for LSTM classification / pairwise models
Summary: Add new LSTM models. This is primarily so we can add compatible pre-trained lstms for similarity. However this is also useful for compatibility with our transformer models (re-uses tensorizers, fits the interface for encoders, re-uses bert / roberta code) and is easier to specify and initialize than the current lstm configs.

Reviewed By: rutyrinott

Differential Revision: D25446726

fbshipit-source-id: a4e8c925e91bd9cbb9dd83458b4fab940381eab8