Vocab should have len()
Summary:
Adding to support len() for the vocab.

Couldn't use     torch.jit.script_method
for the method as it inteferes and len doesn't get called. Using nn.Module for this should fix that in the future.

If I add torch.jit.script_method, the len() function doesn't exist any more.

```
> ======================================================================
> ERROR: test_len (pytext.torchscript.tests.test_vocab.VocabTest)
> ----------------------------------------------------------------------
> Traceback (most recent call last):
>   File "/home/snisarg/fbcode2/fbcode/buck-out/dev/gen/pytext/torchscript/tests/test_vocab#binary,link-tree/pytext/torchscript/tests/test_vocab.py", line 79, in test_len
>     self.assertEqual(len(self.vocab), 5)
> TypeError: object of type 'ScriptVocabulary' has no len()
```

Reviewed By: Darktex

Differential Revision: D25734881

fbshipit-source-id: 10a7df77ea3b93fae19a0607e2e2f6b4eeee855e