Using larger ProcessGroup Timeout to prevent Initialization Errors (#1567)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/1567

The time to initialize each process in PyText is very large so there may be a 40-50 minute gap between the first and last processes being initialized (ProcessGroup being constructed). The previous barrier was not subject to timeouts (since it would block on creating NCCL communicators) so the the initialization would take a long time, but still complete successfully. With the new store-based barrier, there is a busy poll that checks whether each rank has incremented a counter in the store, and if all processes have not joined until the timeout, an exception is thrown (which is the error users are seeing).

Increasing the ProcessGroup timeout in PyText will ensure the store-based barrier does not timeout initialization any more.

Reviewed By: mwu1993

Differential Revision: D25792895

fbshipit-source-id: 4561698fd511d60a5fc00c23418cb6e6d4356787