Script to write pytext config version & pytext fbpkg version mapping
Summary: For pytext xpkg call, we want to cast pytext config to the version that prod pytext fbpkg can take. Right now we depend on manual update to the config, which is not scalable. Write script to automatically update config when new pytext pkg is released.  Will add a node in pytext conveyor as in following diff.

Reviewed By: hudeven

Differential Revision: D26123560

fbshipit-source-id: bea7c43b9f2ca8fed7a1716f4988a3370690814d