Better test for config adapter
Summary:
Starting v23, we start to implement downgrade logic, the upgrade and downgrade logic are not necessarily the same so we have to have separate testing json files.

Making this change so that we have better test coverage on existing and future config versions.

Every future config adapter will need to have two test files, one for upgrade and one for downgrade

Differential Revision: D26899410

fbshipit-source-id: 1a9274e8cd130f26e5b953f476276d066deb57b5