Back out "Fp16 gradients for pytext"
Summary:
Original commit changeset: c053ee43adb0

See my comments on D27178926 (https://github.com/facebookresearch/PyText/commit/44c0e1f0ac3972a030b4832c97bc3d86d7c4c371) ( it broke a test and a pair of integraiton tests for fluent2). It's too old of a diff to ninja unland, so unlanding with this diff.

Hi-pri because this would block fluent2 release on Monday, and we're looking for a couple of bug fixes to land.

Differential Revision: D27223480

fbshipit-source-id: aa3c10d4c42456c3ba727e676c922a258f389608