Summary: This diff aims to mitigate two things:
(1): Discrepancy between config.layer_config.encoder_embed_dim and
config.encoder_config.encoder_embed_dim

Right now there are some duplicates in terms of how the encoder dimension can be defined due to legacy. In this diff one of them is been overwritten to ensure consistency. This is a temporary approach until we have the AssistantEncoder.

(2) There are cases where the input dimension to the encoder is different from the encoder dimension. For example when we use self.pos_embed with a projection dimension.

Reviewed By: AkshatSh

Differential Revision: D27413108

fbshipit-source-id: fa34e9ad478f8a8805855524e8afc52f7be69875