Removing dictionary confidence for NLU.
Summary:
In our NLU models we take a weighted average of non text features (coming from NER, dialog context, etc.) and we rely on both the feature and confidence, we want to understand what the impact is of removing confidence.

We use a new parameter use_weights for DictFeatConfig which is True by default. However setting this feature allows us to measure the impact of eliminating weightage of embedding confidence

Reviewed By: AkshatSh

Differential Revision: D27631609

fbshipit-source-id: 412e699c22a77f41dc426340f301a3361f50b8b3