Fix add_bias_input_layernorm (#1676)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/1676

For layernorm, since we need to do two reductions in the middle, we need to do 3 block-strided loops. But previously, we only record the aggregated value of `local_out`, which was incorrect. Instead, we should record the `local_out` at each loop iteration. This diff reuses the out array to save such local_out, since in the end, out will be overwritten by the finally result anyway.

Differential Revision: D27789994

fbshipit-source-id: 5be3a359163761055ff51e095f795d20b759bf57