Improve input tensorization by capping final sequence at max_seq_len instead of both question and document.
Summary:
RoBERTa/BERT encoders have a max_seq_len of 512 so any input sequence over that needs to be truncated.

But since Q&A tasks take a combination of two sub-sequences, the default behaviour was to apply truncation to each of question and document sub-sequences.

Since 512 cannot work as max_seq_len for truncation per sub-sequence, we used to set it to 384 typically to allow enough room for both question and document. But this makes us lose a lot of document context in the input tensor - 384 tokens is not usually enough to represent larger contexts.

This diff fixes the issue by applying truncation only on the final input sequence so we can see as much context as possible for the encoder.

Reviewed By: stanvp

Differential Revision: D28209117

fbshipit-source-id: 53562296100332d832b404d0ade83f0a031cce09