N-best evaluation in MQT
Summary:
The joint intent-slot model does classification for intent prediction, and per-token classification to predict slot labels or nothing for non-slot tokens. From this model we get a list of scores for each intent, and can then decode slot predictions using only slot scores valid for this intent.
However, in our Model Quality Evaluation (MQT) code we use to evaluate trained models we have a discrepancy, which doesn't affect our top-1 scores, but affects ranking scores, when we want to look beyond top-1 scoring intent.

We want to consider intents beyond top-1 scoring for N-Best evaluation. We underestimate actual performance, when we expect top candidates from the same bucket.

To fix this, we switch from decodeFlatPrediction -> decodeFlatPredictions, and then tracing the code up to the callers to update return types, and make sure we process all the candidates except top-1.

Reviewed By: DenXX

Differential Revision: D30093403

fbshipit-source-id: 62bff2b0ad2b10835414aa7281bd09e2b79f6cd6