Class weight implementation via tensorizer
Summary:
### Background
The current domain classifier is a multi-label classifier with no class weights used. [Previous experiments](https://fb.workplace.com/notes/964904204082889) suggested that growing the size of CQA causes it to be predicted with higher confidence more often, and it starts to be ranked higher than recipe, OOD, and other domains. Moreover, we see even for overlapping cases, the second candidate has lower confidence, often < 0.1. N-Best from NLU should provide a good signal downstream regarding which predictions we consider to be correct.

## Method
We implemented 3 ways to calculate the weights:
1. pos_weight = # negative examples / # positive examples
2. pos_weight = sqrt(neg/pos)
3. pos_weight = # avg positive / # positive if # positive is greater than average, otherwise 1.0

After looking at the domain frequency distribution, a few domains has excessively small number of examples, which makes pos_weight excessively large. This results in sub-optimal results in method 1, therefore we come up with method 2 to normalize the scores. Note that hyperparameter tuning is needed as we change the scale of the loss, especially for method 2). It is not included in this diff and there will be a separate diff on hyperparameter tuning if needed.

The intuition for method 3 is that we only want to penalize the domains that have very large number of examples.

previous diff: D31655884

Reviewed By: DenXX

Differential Revision: D31786137

fbshipit-source-id: 2b2b88a54ffc666dd68b5173abc4ebe7970c02ca