docs: add GH button in support of Ukraine (#1716)
Summary:
Our mission at [Meta Open Source](https://opensource.fb.com/) is to empower communities through open source, and we believe that it means building a welcoming and safe environment for all. As a part of this work, we are adding this banner in support for Ukraine during this crisis.

Pull Request resolved: https://github.com/facebookresearch/pytext/pull/1716

Reviewed By: arbabu123

Differential Revision: D34622563

Pulled By: dmitryvinn-fb

fbshipit-source-id: 44d1b368a40eb3366f7b58fe45e25508db314cdc