Remove fixed pad_shape from FloatListSeqTensorizer
Summary:
Current FloatListSeqTensorizer fixes pad_shape to initial batch. This raises problem in the last batch where number of samples is not sufficient to fill the full pad size. This problem is hidden when BiLSTM representation is used, because of the sorting operation (even with disable_sorting, the index selection part of the method still kicks in and removes the blank slices). When other representation, e.g. CNN is used, this problem shows up.
See run f338219453 for example inputs for BiLSTM representation and run f338248738 for CNN representation.

Differential Revision: D35598633

fbshipit-source-id: 9820fcb13da7ea7002d5264439d8fa15bf3fdb7b